requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
crewai==0.11.2
langchain==0.3.0
langchain-openai==0.2.0
python-multipart==0.0.12
websockets==13.1
pydantic==2.10.0
sqlalchemy==2.0.36
alembic==1.14.0
redis==5.2.0
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.0
aiofiles==24.1.0
Pillow==11.0.0
speechrecognition==3.10.0
pydub==0.25.1
openai==1.54.0
anthropic==0.40.0
requests==2.32.3 